Dev

See merge request consultadd_products/log1!199

+++ project/models.py
@@ -277,7 +277,7 @@ class ConsultantFeedback(TimeStampedModel):
 class ConsultantLeave(TimeStampedModel):
     year = models.IntegerField(_('Year'))
     balance = models.FloatField(_('Balance Left'))
-    granted = models.IntegerField(_('Granted Leaves'))
+    granted = models.FloatField(_('Granted Leaves'))
     is_expired = models.BooleanField(_('Leave Expired'), default=False)
     leave_type = models.ForeignKey(Choice, on_delete=models.CASCADE, related_name='type')
     consultant = models.ForeignKey(
